pyproject.toml
[tool.poetry]
name = "pgbackup"
version = "0.1.0"
description = "A cli program to automate your postgresql databases backups"
authors = ["Tobi-De <tobidegnon@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
typer = "^0.3.2"
boto3 = "^1.18.1"
psycopg2-binary = "^2.9.1"
pytest = "^6.2.4"
emails = "^0.6"
pydantic = {extras = ["email"], version = "^1.8.2"}
schedule = "^1.1.0"
pytz = "^2021.1"
APScheduler = "^3.7.0"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
black = "^21.6b0"
isort = "^5.9.1"
typer-cli = "^0.0.12"
bpython = "^0.21"
python-dotenv = "^0.18.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
